mem-stack: fix performance issue in mem_stack_protect under ASan.
mem_stack_protect is called by mem_stack_pop in order to ensure that using
mem-stack allocated memory after a frame has been popped leads to an
invalid read. Previously, the implementation traversed from the end of the
active frame to the end of the last block. If only a few bytes of the last
blocks were used by the popped frame, this is a waste of time (even more
in case the blocks are huge).

We now only protect from the end of the active frame to the end of the
popped frame, ensuring we protect only what is needed.

Change-Id: I325cbeab1f405252592abf1a2c078f23359b6671
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>