core-mem: rework allocators.
This patch does an important rework of our allocators in order to get rid
of some historical limitations.

First, the mem_pool_t structure becomes the corner stone of our allocation
scheme instead of the MEM_* flags. As a consequence all the allocation
helpers are implemented using a variant that uses a mem_pool_t, then
specialized version for different pools are provided.

Secondly, the libc allocator is now a properly defined mem_pool_t and not
a 'debug' pool left alone in a core-mem-debug file. In the deeper layers
however, NULL is treated as an alias for libc, providing proper support
for what was already used as a convention in most of our allocators. The
mem_pool_t also gains a pool member that contains the old MEM_LIBC value
(in case of the libc allocator).

Thirdly, the alignement constraint of the allocator is now made available
everywhere. It is part of the base API of the allocators (note that is is
an API breakage). For now only the libc and the stack allocators support
alignement constraints.

The main consequences are:
* API for mp-based allocators is much more complete.
* low-level allocators are mp_imalloc(), mp_irealloc(), mp_ifree(), all of
  them supporting NULL to specify the libc allocator.
* old imalloc(), irealloc(), ifree() are just helpers to call the mp_*
  version with the right pool.

More work will come in order to improves support for "other" mem-pools in
our base libraries.

Change-Id: Ib2ce7f781b8ded1a06927e744608db8986e91282
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>