mem-stack: support alignments smaller than 16 bytes.
The mem-stack used to have a minimum alignment of __BIGGEST_ALIGNMENT__
which is 16 on our platforms, which is a waste of memory since most of
the time an alignment of 4 or 8 bytes is strictly enough.

The change allow alignments as small as 1 byte in release mode but is
constrained to 8 bytes in development mode (in order to maintain the red
zones between the allocations). Also, __BIGGEST_ALIGNMENT__ is used if,
and only if, the user didn't provide any alignment hint, which is now
really a corner case since most allocation helpers use the alignof() of
the allocated type.

Also fix some wrong size in reallocations: the alignment of a reallocation
must match the one of the allocation.

Change-Id: If03da10f3463ac4e040e13153615922cddd965c5
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>