mem: put the bit-alignment primitive at top level.
That way it is accessible from all the allocators and not only from
mem-stack. The bit-alignement use the log2() of the alignement in order to
use very-fast arithmetic to perform pointer alignment instead of a
ROUND_UP() that would use a division.

A side effect of this patch is that mp->malloc() and mp->realloc() MUST
NOT be used directly, the right API to use are the helpers mp_malloc() and
mp_realloc().

Change-Id: I3649e692e1e226cc346ca054d45a415045c9c723
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>