lib-common: Port ic and thr to use core-module API
The goal of this commit is to make both ic module and thr compatible
with the core-module API

For thr module:
  Problems: fork() and thr

  Example:
    Z_GROUP_EXPORT(qps){
       thr_initialize();
       Z_TEST(qps1) ...
       Z_TEST(qps2) ...
    }
        z_run-> |
                | <- thr_initialize() => REQUIRE(thr) => thr constructor
                |
         fork-> |
                -----------
                |         | <- thr_atfork()     => wipe all threads
                |         |
                |         | <- thr_initialize() => REQUIRE(thr)
                |         |                     => Do nothing since the module is loaded
                |         | <- qps segfault

The problem is that thr_atfork() is deleting the threads from the
forked one, thus disabling the thr module but without telling the
core-module. So on the next call to thr_initialize, the constructor
won't be called because core-module thinks it is already loaded.

The solution of calling thr_initialize() using the pthread_atfork()
function don't work since when the thr_atfork() function is called the
process is in a state where there are 10 open threads.

The solution is to not call directly fork() but use thr_fork() instead.

The function thr_fork() forks the process, but do some additional
process for the child.

pthread_atfork is still used for backward compatibilty but should be
removed every fork has been changed

Change-Id: I5f9f45a42edfea2bfe62e303b25ffb388c99b491