darwin: add compatibility layer.
This adds the compatibility files to support darwin as a target OS. This
patch mostly redefine functions that are GNU specific and thus absent
from the FreeBSD setup

This patch touches unix.blk because the *at() function need the function
fd_get_path() that is defined there.

Note that mremap is *not* defined.

Change-Id: I9ab715ce4c0ca79ad0d905983b34af44bed686d5
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>