lib-common: fix day saving time in localtime_cur/next*
The issue was for calculating period during the time change.
The calculation was wrong due to the auto correcting functionality
of mktime.

As example, in 2013 the time change From CEST to CET was October the 27th.
So asking for current week of October the 27th at 12:00:00 (CET)
was supposed to return October the 21st at 00:00:00 (CEST).
However the returned value was October the 21st at 01:00:00 (CEST).

This was because in the tm structure the isdst value was not reset.
So mktime was called with the incorrect time October the 21st at
00:00:00 (CET), which was auto translated into CEST.

Change-Id: I5586afc61900062fdadcf2fad4ca788ca08df499
Signed-off-by: Sebastien Boulle <sebastien.boulle@intersec.com>