qhash: add qm_get() and qm_get_def() helpers.
Those helpers perform qm_find() + values[pos]. They will let us simplify a
lot of cases where accessors to the qms are overloaded with a find
function that does the exact same qm_find() + values[pos].

We have two variants:
* a qm_get() that will crash if the entry is not in the qm.
* a qm_get_def() that will return the default value if the entry is not in
  the qm

Note that these functions are not here to replace qm_find that will still
be the most generic approach when trying to perform complex lookup
operations. Moreover, no function that return a pointer to the value is
provided in order to avoid qm_get* to be used as qm_replace()
replacement (which would be insane).

Change-Id: I5a96b306df191ceb5bef3f51f02a869eb769d77d
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>