iop: add IOP_CLASS_SWITCH() and IOP_CLASS_EXACT_SWITCH() macros.
Those macros will let you switch on the type of a class. The EXACT variant
matches only cases that have the exact same type as the class instance,
while the non-EXACT variant matches the nearest parent of the class.

Bump the iopc at the same time, this will be useful.

Change-Id: I57cab5c555dfe6fc2dc15a3ab96bc31c7c1b01ce
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>