Build: fix syntastic target.
since commit 45f9e93b00ff of syntastic, we need to have one argument per
line in the syntastic config file (this is also documented in
https://github.com/scrooloose/syntastic/wiki/C%3A---gcc)

There is also a problem with arguments using quotes. For example
-D"index(s,c)=index__(s,c)".
Older versions only work with the quotes and newer versions only work
without them, so I chose to just strip the option.

Change-Id: I7b98ea3ef4a9a6ad23b602e88a1c943c1a1d212b