z: add ability to specify a directory for temporary data.
By default z use the root directory of the repository as its root path for
temporary directories. However, in some circonstances (on some machines),
that directory may not be appropriate to run the tests.

We now allow the definition of a z-directory that will be used in
priority. This directory can be specified either using an environment
variable (Z_DIR) or via command line (--z-dir).

Change-Id: I1a7b067860b84ec63f79bbc98f50ba7d3021e6cc
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>