module: improves compile-time validation.
In order to improves compile-time checks, we now have a module_t * object
made publicly available in the headers of the modules that declare a
module. The core-module internal API has been updated to use the module_t
* as input instead of the names, which ensure that the module actually
* exists at required/release time.

The sole remaining issue is for dependences that are still referenced by
name because its hard to do otherwise with macros, and also because of the
randomness of the registration procedure.

Another indirect improvement is that this will avoid having to use the .wa
everytime we use a module since the modules are now explicitly at
link-time by a side-effect of using public variables.

Change-Id: Iffd00d42ce11a5b1aaf8eb5279503b43cc8c2693
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>