iop inheritance: copy an object into a derived one
Append a function to copy an iop object into another one which is of same
class type or of a child class type.

The goal is too put the values of the parent object into the child one
without modifying the __vptr of the child.

Flags allow duplicating of the data. Data duplicating is not yet
implemented.

Change-Id: I99de5f4df8533252a348c64e03a7174e5267876d