logger: logger now use module
Problem:
-------
There was a race between destructors.
logger_shutdown was called before module_shutdown, thus when core-module
was closing modules, if one of these modules tried to write some logs
it would generate a segmentation fault.

Solution:
--------
* Logger now use modules, and we make every modules which register with
MODULE_REGISTER depends on the "log" module.
That way, we are sure that we can use logger during the shutdown
procedure of our modules.

* Also in hard shutdown we no longer emit a trace when a automatic
required module is not shutdown. At this point the logger module is
normally shutdown.

* the module "thr_hooks" need to be register manually, in order not to
  depends on "log". (races during constructor)

Change-Id: I8ba8450111bdf56c1de5d1e57f21be61655ccf8a