thr: fix thr-job reload on fork.
We had an issue that caused thr-jobs to be wiped as many times as we
forked since we registered the pthread_atfork callbacks in the initialize
that is called after forking.

We fix it by ensuring pthread_atfork is called a single time: the
registration is now performed at the same time as the module registration
in a constructor function instead of being in the initializer that is
called after every fork.

As a consequence, there's no need for a thr_fork() anymore since we have a
better control on when thr_initialize() is reloaded than before.

Change-Id: I0337a68f7e8165eac028ddfcac34c57ef1d7c62e
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>