Fix build with clang 3.4.
clang 3.4 now detects inline function that are in the main file but are
unused and emit a warning for that. This allows detecting more dead code
than before (yeah!) however this emits a warning for almost every single
function generated by the qvector_t() or qm_*_t().

It also detects duplicated "inline" attributes.

Change-Id: Ie68bd6343346b7d1c117d4571396e0470aa539e9
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>