core-module: add a new flexible module definition scheme.
The use of the macros MODULE_REGISTER() and MODULE() is quite limiting due
to the fact we want to include a lot of information in a single macro. To
make module definition more flexible, we now introduce the new macros
MODULE_BEGIN()/MODULE_END() that look a lot like the
OBJECT_VTABLE_BEGIN()/OBJECT_VTABLE_END() of the core-object module.

The main improvement for now is that MODULE_DEPENDS_ON() macro that can be
used in a module definition block have a real object-dependence on the
target module, forcing it to be loaded when the current module is loaded.

A typical block will look like:

 MODULE_BEGIN(toto, &toto_on_term)
     MODULE_DEPENDS_ON(titi);
 MODULE_END();

This is much more verbose than the previous API, but it will allow having
a more stable API in the long term.

Change-Id: I94d34704b9662aedd204e1e1d8f0db0b2b71ad94
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>