core-module: add the notion of "methods"
A method is a callback that can be optionally implemented by a module and
that is called when an event occurs.

The method is identified by an object that can be defined using
MODULE_METHOD_DECLARE/MODULE_METHOD macro. This takes a name and the
prototype of the method (methods can either take no argument, take a
pointer, an integer or a data_t object).

A module that implement a method must include a call to the
MODULE_IMPLEMENTS(_TYPE)() macro in its module definition block.

  MODULE_METHOD(INT, DEPS_BEFORE, my_method);

  MODULE_BEGIN(toto, NULL)
      MODULE_IMPLEMENTS_INT(my_method, &toto_my_method);
  MODULE_END()

Then, the method can be called using the module_run_method() function that
takes the method descriptor and the argument to be passed. The method is
then run on all the loaded modules that implement it in either a
depth-first or breadth-first fashion, depending on the configuration of
the method.

Change-Id: Ib55521530a8b2f1ab8c236a351e97ef889b62bd1
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>