core-module: fix crash when loading some .so.
When loading a dynamic library linked against the lib-common, as it is the
case in the teaser, some module get registered more than once. Workaround
this by returning the existing reference in case a module is registered a
second time.

This fix the crash at the startup of the teaser.

Change-Id: I909aa044afb0e91ce99b30d85df73d38bab5a619
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>