lib-common: fix version script
The command 'git describe' searches through the tree differently
if the match option is active or not.

Thus in the branch origin/lms/5.1/maint, on the commit 5c169deb1,
the following behavior is observed:

git describe
> lms/5.0.8-4599-g5c169de

git describe --match="lms/*"
> lms/5.1.2-93-g5c169de

The first-parent option is useful when you wish to not match tags on
branches merged in the history of the target commit.

Change-Id: I9d14470ba0ccab66673d272306fd6bfbf11f2ad3