log: make destructors aware of a critical state.
The goal of this commit is to create a global variable that allow us to
know when a logger_panic or logger_fatal has been called.

It can be used in the destructors in order to skip some code when we are
in a critical state (logger_panic & logger fatal).

In core-module, we choose to not closed the modules when we are in a
critical state.

Change-Id: Ifb7a738e0ad7baae4cf1cc88d11900eb4883d441