lib-common: fix htlist_deep_clear.
Currently, when clearing a htlist with htlist_deep_clear, list->head was
not set to NULL, so it continues to reference a no more valid pointer. So
we are in a configuration where *list->tail=list->HEAD=??? (let's call it
PTR).

This may be problematic, since we can fill the list again, with a node that
may have this same pointer PTR. With such a configuration, htlist_for_each
is buggy since it compares node with *list->tail, and skips the for loop
without any iteration.

The fix is to set htlist->head to NULL in htlist_deep_clear, this is
coherent to what we do in htlist_init.

A non regression test is added to summarize this scenario.

Change-Id: Ib3e8cedab0bf80c60b74f78c6bdb203f014aa7fa
Refs: #20271