core-obj: support allocation on custom memory pool.
Objects now contain ->mp field that should be used to allocate fields that
are owned by the object (preferably, but not mandatory). This ->mp is the
pool on which the object itself is allocated.

New helpers are added to allocate objects on a specific pool.

This is mostly suitable for pools that have an explicit deletion operation
since other pools don't guarantee the wipe helper is called.

Change-Id: I9ef92f048487e53e401050b7331628b7874e7927
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>