make <log> modules depends on <thr_hooks> module
The <log> module uses some thr_hooks() to work so we have to add
a dependency of the <thr_hooks> module on the <log> module. And since we
cannot rely on the “constructors” initialization order we have to register
the <thr_hooks> module in log.c…

It's not very elegant but since all module implicitly depend on the <log>
module it's not that shocking.

Change-Id: Ifec7630bbddebc454a835c49bcbd956f4e2ea36f