core: introduce expect() macro.
expect() is an expression that returns the value of the condition it
takes. In development mode, the macro behaves like assert() (but it
returns its value), but contrary to assert(), expect() will still be
evaluated in release mode.

The purpose is to make error handling easier by allowing construction of
the form:

 if (expect(pos >= 0)) {
     /* pos >= 0 */
 } else {
     /* pos < 0 */
 }

instead of

 assert (pos >= 0);
 if (pos >= 0) {
     ...;
 } else {
     ...;
 }

or

 if (pos < 0) {
     assert (false);
 }

The returned value of expect() must be checked forcing user to write the
if condition or explicity ignoring it.

Change-Id: I4eaa242901337a3da94e7df2f84fb1125e4c0def
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>