el: ensure children won't alter parent's event loop.
The epoll fd is inherited in the child process, as a consequence every
el_fd_register/unregister in the child affects the parent process as well.
This has been observable twice during the last weeks:
* first, the log fd being shared by the parent and the children, if the
  child did log something, it led to a corruption of the stated of the log
  fd in the parent process: the fd was set in POLLINOUT by the child in
  the epoll context, while the parent still though it was in POLLIN only.
  This led to fea3fa3e65 quick fix.
* secondly, the eventfd used to wake the main thread by the thr jobs was
  being removed from the el-loop by the child, disabling the notifications
  for both the child and the parent process. As a consequence, queued
  event were processed only when the event loop started a new loop.

We fix this by forcing the child process to use a different epoll fd. As a
consequence no fd registered by the parent is watch by the event loop of
the child. This changes the current behavior and my break some products.

This patch is made on master for validation, and will be remade on earlier
branches once it has been checked that it does not introduce any
regression to products.

Cherry-picked from 99e07a8f33164b8796379e2a9.

Closes: #20554
Refs: #20248
Change-Id: I7925ecc63f9576263f92ab7bc2e6de73ed30197f
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>