module: Provide argument in module definition
The goal of this patch is to be able to provide argument to a module in
the module definition (between MODULE_BEGIN & MODULE_END).

There was a problem that the constructor of the module we are providing
the argument has not yet been called.

If it is the case, we detect it and we store in a qm the argument for
the module.

On module registration we check in the qm if an argument has already
been provided.

Example:

MODULE_BEGIN(master_list)
    static platform_list_module_cfg_t list_cfg = {
        .master_ic      = NULL,
        .master_ic_impl = NULL,
    };

    MODULE_PROVIDE(platform_list, &list_cfg);
    MODULE_DEPENDS_ON(platform_list)
MODULE_END();

Closes: #20646
Cc-Redmine: yes
Change-Id: I3621b88b1404c8fee9a88d4cc787324aa44ff3f0