disable atfork handler when building a shared library.
The pthread library does not provide any unregistration mechanism, as a
consequence when a shared library performs a pthread_atfork() registration
and is later unloaded, this causes fork() to call function that does not
exist anymore in the address space of the program, leading to an immediate
crash (I'm quite sure this also could be easily exploitable to inject
random code).

To avoid this, we introduce a define when building files to be included in
shared libraries, and disable the pthread_atfork() registrations in that
case.

Closes: #18218
Change-Id: I5360ef4a20cc45d272daaa4b6c5bcb73370990ce
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>