sb: Add sb_add_csvescape and variants.
We need this function to fix the bug described in ticket #20730.

A function also called sb_add_csvescape was already introduced in
14bc30e4b, but does not have the same behavior.
It is currently available in lib-common 2013.1, 2013.3 and master, but is
not used anywhere.

It will be replaced by this one (which is a priori more useful) during the
merge cascade.

Refs: #20730
Signed-off-by: Aurélien Lajoie <aurelien.lajoie@intersec.com>
Change-Id: I29a538d9a9f9efa02f47367bc07bfe5f0566baaa