module: fix method execution in reverse order.
When the method was called within an initializer in reverse order, the
dependent module is already added in the required_by list of the
dependency, but not started yet. As a consequence, when running method on
dependent first, we need to ensure the module has been started before
calling the method on it.

Change-Id: If07ec9e745a4ac2a3d8070bf3cbe9f57f3967f2f
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>