thr-job: do not relaunch thr-jobs after fork by default
The aim of this patch is to avoid a lot of possible races when forking.
Some tools like Asan don't like much to relaunch the thread jobs after
a fork.

Now people which really want to use threads-jobs after a fork (like QDB)
will have to write something like:

foo() {
    bool prev_state = thr_job_reload_at_fork(true);
    pid_t pid = fork();

    thr_job_reload_at_fork(prev_state);
    if (pid…) {
    }
    …
}

Refs: #20489
Change-Id: Ib7e2e1228f8ce4307687397d609651ca428c14cb