core-module: Be able to disable module registration.
This commit provides a way to disable a module registration declared with
MODULE_BEGIN/MODULE_END if, for example, we are compiling a shared
library.
This is a horrible hack which consists in replacing the
__attribute__((constructor)) by a __attribute__((unused)).

Also use this to disable the modules log, thr and el in shared libraries.
This fixes the teaser bug describe in #21050 (it was not started anymore)
and also removes the annoying warnings "XXX thr has already been
register" (BTW, fix a typo).

Closes: #21050
Change-Id: Ibc79218e0dddde80ed6874910312fe7daf59a464