hash: remove murmur3 x64 32bits.
That function was an improvisation from Pierre since it does not exists
in in the official hash description for murmur. It returned the first
32-bits word of the 128bits hash variant. However, by testing that hash
function with SMHasher[1] behaves dramatically worse that murmur3 x86
32bits (completely failing some of the tests). Moreover, on small key size
(tested with keys as long as 32 bytes), the x86 version performs
consistently (far) better (20% better on 32 bytes, 50% better on 1 byte).

As a consequence, just get rid of the buggy function and uses the x86
variant as our default hashing function.

[1] https://code.google.com/p/smhasher/

Change-Id: I6c7d71f177d6e266eaa1a38a695502e7d11191fb
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>