Build: hide symbols by default.
Previously, symbols were hidden only in release builds (which sometimes
caused issues because some code failed strangely in release while worked
perfectly in development mode). This caused all symbols to be public in
development mode. As a consequence, when loading several time the
lib-common (for example), the dynamic linker could use both the
shared-library version or the binary version of the lib-common, both being
accessing from the shared library. This caused modules of the shared
library to be registered in the binary's core-module, leading to module
redefinitions and ultimately to crashes (as seen on the teaser).

Now, the symbols are accessibles only in the current object, as a
consequence, the log module of the shared library registers itself in the
core-module of the same library since the core-module of the binary is not
visible anymore from the library. This also mean that all the symbols will
be duplicated (for example, configuring the logging module in the binary
will have no impact on the logging module of the shared library).

Note all this only affect development builds, release builds were already
acting that way.

Change-Id: I9c0a90f1c54099a7d0af8e20bd37e37b5080ee9c
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>