qvector: add qv_(deep_)extend helpers
We could have some case where we want to append the content of a qvector
inside another one. So we add two helpers inside lib-common to handle
it.
The first one copy (without duplicate it) the content of the source at
the end of the destination, and the second one take a block to duplicate
the source into the destination.

Closes: #21206
Change-Id: I3bc98e0123b7be487e857fe17fbc5ea8de953686