mem: add more information in mem_pool.
The mem_pool structure now describes how it works using a set of flags:
- MEM_EFFICIENT_REALLOC if it can realloc efficiently (with little risk of
  copy)
- MEM_BY_FRAME if the allocations are grouped and cannot be deallocated
  independently from the frame

Also, mem-pools can now have a fallback allocator that is used in case of
a reallocation of an pointer of that pool. The associated
mp_irealloc_fallback is provided and perform the allocation on the fallback
pool if needed.

Change-Id: Iaba5ef4bf5e5b1d78f53be9701af09345a14ce98
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>