qvector: fix buggy free in some cases.
The skip field of the vector is expressed in number of elements, not in
bytes. As a consequence, it must be multiplied by the size of the entry
when computing the base pointer of the allocation.

There were two cases where we did free the wrong pointer however: when the
vector grows while containing skipped entries, and when qv_optimize() is
called.

By chance, skipped items in qvector are not really commonplace.

Change-Id: I08c86a5e93215eb0892b120a5a151b9edb56f8bd
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>