str-buf/bit-buf: switch to mem-pools.
A sb (or a bb) can now be allocated on the mem pool of your choice, and
will be reallocated there. Two modes of initialization are still allowed:
* by using an empty sb that points to __sb_slop
* by using a pre-allocated size

The second solution is well suited for custom allocation pools, and in
particular in case the pool does not provide an efficient reallocation.

This patch is dangerous since it changes the behavior of __sb_slop that
previously was writable but isn't anymore.

Change-Id: Ia1cb4375d08559f2a994ffd243e2694d0e4dbb3f
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>