ichannel: secure the “owner” property usage
In the patch f96cfc4 we tried to explicitly forbade manual deletion of
IChannels using the “owner” property (reserved for ic_spawn() & autodel
usage). But it seems that too much code uses this pattern…

So we fix ic_delete() to always delete the original pointer even if
ic_wipe() put the “owner” to NULL first.

Change-Id: I568abbb09be018d5284d8c1db4b80f446bc3289e