core-mem/tlsf: fix broken function casts
The casts were introduced in mmsx, patch d531c52d082a13446 for some
unknown reason.

It appears that the casted functions matched the target prototypes at the
time and still did until a rework in the memory pools in lib-common
2014.1, patch e45c9aeb. This is the reason why this patch is not proposed
for an earlier version of the lib-common: we don't expect this
compatibility to be broken again in any earlier version.

The forced use of incompatible functions pointers led to the corruption of
the memory flags when calling tlsf_malloc() or tlsf_realloc() through the
memory pool interface.

Closes: #22267
Change-Id: Ic696e4f6282e9e34efc78461c2102f3c73f75e41
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>