log: fix use-after-free.
When emitting a log while a frame deletion is pending, that line of log
wasn't put in the correct frame of the mp_stack. For example:

 start() -> push frame 1
 start() -> push frame 2
 stop()  -> return frame 2
 log() -> allocate in frame 2
 stop() -> pop frame 2, return frame 1

In order to fix the issue, we ensure that the frame is poped before
logging the new line:

 start() -> push frame 1
 start() -> push frame 2
 stop()  -> return frame 2
 log() -> pop frame 2, allocate in frame 1
 stop() -> return frame 1

Note however that this means that using using the returned log_buffer
cannot be mixed with the use of log function in case you have several
buffering frames. This dramatically reduce the usability of the imbricated
buffering. Some additional thoughts will be needed in order to make this
feature both correct in term of memory management and usable.

Change-Id: I108c05f9bf9c6350d53f0bcf193f2a21442ac4ca
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>