iop: add 'type' field to iop_static_field_t struct
If we want to read the value of a static field without knowing its type,
there should be a type field inside the iop_static_field_t struct.

For backward compatibility, add the new field at the end of the structure
and add a struct flag noticing the availability of the new field.

It will be used in tools for iopy class generation to read the static
fields.

Change-Id: I62006f384b1bf908d01a98cad08b4e0b8e67f618
Refs: #21872, #22257