Build: add minify support for js files
In production, we need to have compressed js files.
The r.js tool supports the compression with uglifyjs but
this feature is not very extensible.

This patch adds the uglifyjs support to compress the
generated files from r.js. This implementation depends
on r.js target. So currently, we need to initialize r.js
to use MINIFY target.

 _JS = blah

 blah_SOURCES = htdocs/javascript htdocs/iop
 blah_CONFIG  = build.js
 blah_MINIFY  = htdocs/javascript-compressed/pages/login.js

Change-Id: I6857eb3362c8955b2ea81f777c6dc6795e418694