core-mem: Use an enum for mem pool flags.
The problem with the previous defines is that MEM_STATIC (which is equal
to 0) can be given with no compiler warning to any function taking a
pointer on something (and is interpreted as NULL).

This is very problematic since the recent rework of sbs (f9f5c57e) in
which the last parameter of sb_init_full became a mem_pool_t *: the calls
to sb_init_full giving MEM_STATIC as argument were still compiling but
became buggy (segfault at first reallocation because the content of the sb
was not malloced).

Using an enum instead of defines allows to have compiler warnings (with
clang 3.4 only).

And obviously fix the buggy calls to sb_init_full.

Change-Id: If9a70e020098dc8289578e1f8ee9e3e56279afd8