qhash: protects against overflow when looking in header bits
The qhash_slot_is_set() function didn't check its boundaries when looking
for a bit. It was a problem in the qhash_del_at() function because the
given position could belongs to the “old->bits” array. And this array
can be larger than the “hdr->bits” array (which is accessed first).

This situation only happened if the hashtable was shrunk (a lot of
elements removed then a new element added). But only if:

- the qh was sparse (to keep large keys after rehashing);
- the qh gets really shrink and with the broken condition
  "if (newsize < hdr->size / 4) { newsize = hdr->size }" it's quite hard
  to obtain;
- the qh was big-enough to not remove its “old“ bits in one pass.

Unit-tests will be added in more recent lib-common and will be:

Z_TEST(shrink_bug , "test for hdr->bits overflow after shrink") {
    qh_t(u32) qh;

    qh_init(u32, &qh, false);

    qh_set_minsize(u32, &qh, 1);
    Z_ASSERT_N(qh_add(u32, &qh, 0));
    Z_ASSERT_EQ(qh.hdr.size, 11U);
    /* Force qh.hdr.size to grow up to x16 */
    for (int i = 1; i < 8192; i++) {
        Z_ASSERT_N(qh_add(u32, &qh, i));
    }
    Z_ASSERT_EQ(qh.hdr.size, 12289U);

    /* Remove some elements */
    for (int i = 1; i < 8191; i++) {
        if (i % 42 == 0)
            continue;
        qh_del_key(u32, &qh, i);
    }
    Z_ASSERT_EQ(qh.hdr.len, 197U);
    Z_ASSERT_EQ(qh.hdr.size, 12289U);

    /* Now add one element which should force a shrink */
    Z_ASSERT_N(qh_add(u32, &qh, 1));
    Z_ASSERT_EQ(qh.hdr.len, 198U);
    Z_ASSERT_EQ(qh.hdr.size, 3079U);
    Z_ASSERT_P(qh.old);

    qh_for_each_pos(u32, pos, &qh) {
        qh_del_at(u32, &qh, pos);
    }
    Z_ASSERT_EQ(qh.hdr.len, 0U);

    qh_wipe(u32, &qh);
} Z_TEST_END

Change-Id: I89d855e1fc27a5fc9d174e21427510a3ecb6dc2e
Closes: #22613