lstr: Set mem_pool to real one when possible.
In all the functions taking a mem_pool_t *, this is possible to know the
exact mem_pool flags instead of choosing MEM_OTHER.

This was not a bug, but like that the wipe is optimized for some memory
pools.

Also modify mp_lstr_wipe and _lstr_copy to make them smarter.

Change-Id: Ideea879043b6b5ad83d9d5fcfd65390b68e541d7