str-l: make mem-pool based API a first-class citizen.
The allocation API of the lstr was quite inconsistent in that it didn't
contains the same functions from one mem_pool to another (for example,
there was t_lstr_cat() but no lstr_cat()). First this by having a first
class mp_lstr_*() API and then implements the heap and t_stack variants
based on this. Any allocation possible on lstr_t is now possible on any
mem_pool_t.

At the same time, get rid of some redundant mem_pool_t and mem_flags in
some function calls and reorganize and fold the whole file.

This creates a huge diff with very limited actual code change but some
reordering was required in order to have called function defined before
callers, so the diff was required anyway.

Change-Id: Ic2d27e110a91faa2095198e18493e134cf7afd17
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>