lib-common/parseopt: add OPT_UINT and manage 64 bits integer.
When using the parseopt function, it was only possible to parse flag, int,
str and char. The result of the OPT_INT was an int, which makes impossible to
get an int64 or an uint64.

This patch adds the OPT_UINT option.
And save the size of the int to give back an 64, 32, 16 or 8 bits (unsigned)
integers.

Change-Id: If3ef51b1d1e550761bd0954679361ac698882464
Signed-off-by: Florent Tribouilloy <florent.tribouilloy@intersec.com>
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>