atomics: replace custom atomic operations by C11 standard ones.
C11 comes with a new standardized set of atomic operations defined in
stdatomic.h. This come with a sane memory model that is both properly
defined and far superior to the one used in our custom core-atomic
helpers.

Remove the old core-atomic stuff and import a free implementation of
stdatomic.h (taken from github[1], itself adapted from freebsd, that support
both GCC and Clang). The spinlock type has been moved to core-types.h.
Port all the code to the use of the atomic types and helpers without much
thinking for now.

The most noticeable changes are:
* the CAS prototype is a bit different: the stdatomic version feels the
 "expected" pointer with the stored version in case of failed comparison.
* all accesses to "atomic" variables are wrapped at least in
 "atomic_load/atomic_store" macros. We use relaxed order everywhere the
 a direct access to the memory was previously used (i.e. without atomic or
 shared macros)
* barriers are all replaced by acquire+release fences for now.
* thr_syn_wipe() use a proper acquire semantic on the refcount.

[1] https://gist.github.com/nhatminhle/5181506

Change-Id: Ia91c4b171fe6af9862c6c1707693357c292f2635
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>