core: use C11 _Static_assert() when available.
_Static_assert() is available in compilers that support C11 features. On
GCC it appears to be treated as a declaration statement, which forbid its
usage in the middle of a block (because we forbid mixed declarations and
statements), that's why the STATIC_ASSERT() macro wraps the
_Static_assert() statement in a do {} while (0)

Additionally, __error__(msg) is also defined as an empty macro in that
case. It could have been implemented as a _Static_assert(false, msg),
however this breaks asn1-macros that relies on the compiler dead-code
elimination pass to detect unsupported fields but that pass is executed
after the evaluation of the static assertions.

Change-Id: I07a0fefb0698d159c313c69e8189e582d928c5de
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>