thr-jobs: fix data-race between job enqueueing and job stealing phases.
The race was reported by TSan, it was a consequence of the fact there is
no real synchronisation point on deque_entry structure. The stealing
procedure used to use the stealing by first reading the entry then taking
the ownership using a CAS. However, there was no explicit synchronization
on the entry.

The new procedure now takes the ownership of the entry first, and then
marks the entry as used by removing the pointer to the job that became the
synchronization mark.

Change-Id: I591676fdbfb415c2cc67cb0e4b605629afc85de5
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>