str-buf: fix write to __sb_slop in sb_addvf.
As for previous patch, writing the sb_slop buffer is a bug, so ensure this
won't happen. In order to achieve this, consider that if the sb contains
no available space, then there is no need to try to write in it at all:
when there is no available space, just give 0 to snprintf().

Also avoid double formating if the buffer had the exact right size.

Spotted by TSan.

Change-Id: I1460da6752c5f0afc5ed058e528628a71fb9ba40
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>