thr-jobs: disable tsan instrumentation in thr_wipe.
TSan does not know about fork(), which is quite a shame because this means
it does not know the other threads are dead after a fork. As a
consequence, thr_wipe() reports data races with dead threads on
thr_parallelism_g.

With this patch, lib-common produce no data-race warning when running
make P=tsan check.

Change-Id: Ia11b2cf241e5c1562cbef08cdc6926e604223987
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>