ASan: get rid of the handle_segv=0 requirement with QPS.
We know enforce allow_user_segv_handler=1 flag so that the sighandler we
set up are not discarded by ASan and we chain our segfault handler with
the previously set one (with by construction will always be the one from
ASan).

Also, disable panic handler registration so that we always end up in ASan
signal handler.

This may somehow alter the behavior of the programs in case of crash if
handle_segv=0 is specified, however that parameter will not be required
anymore (no ASAN_OPTIONS will be required at all to run our products).

Change-Id: Ib015db95d6f63fe974c267c025c5183d5ab3ef55
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>