licence: introduce modules
The modules/features need to have different expiration dates, the use case
is "try before you buy".

The Licence contains an array of LicenceModule, then each module can have
its own expiration date. There is an API in order to declare the modules
which are used by a product, and another one to determine if a module is
activated or not, giving a licence.

Refs: #23849
Change-Id: I7bb68d83e9f23081aed795d77d0e2b300a97abfb